% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{journal}
\alias{journal}
\title{Titles and abstracts of articles from the Open Journal of Statistics}
\format{A data frame with 10 rows and 2 variables:
\describe{
  \item{Title}{Title of article}
  \item{Abstract}{The first sentence of the abstract}
}}
\source{
\url{http://www.scirp.org/journal/ojs/}
}
\usage{
journal
}
\description{
A dataset contain the titles and the first sentence of the abstract
for 10 different articles in the Open Journal of Statistics.
}
\keyword{datasets}

