% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kwmatch.R
\name{keywordMatch}
\alias{keywordMatch}
\title{Match by Keyword}
\usage{
keywordMatch(x, variation, one.to.one = FALSE, rm.punctuation = TRUE,
  rm.stopWords = TRUE)
}
\arguments{
\item{x}{A column of phrases, which may correspond to multiple elements in the variation column \code{x}}

\item{variation}{Contains phrases which are compared to phrases in x \code{variation}}

\item{one.to.one}{If TRUE, keywordMatch attempts to match each phrase in x
to a unique phrase in variation, otherwise each phrase in variation is matched
to is most similar element in x \code{one.to.one}}

\item{rm.punctuation}{If TRUE, remove the punctation from all phrases \code{rm.punctuation}}

\item{rm.stopWords}{If TRUE, remove stop words from all phrases \code{rm.stopWords}}
}
\value{
correctly.matched A data frame with 2 columns where phrases from x are
lined up with their matched phrases in variation. Pairs that have the highest
cosine similarity are listed first.
}
\description{
Matches a set of phrases in one column to its closest corresponding
phrase in another column, utilizing the cosine similarity based on the
tf - idf statistic.
}
\examples{
set.seed(100)
journal <- load(kwmatch.journal)
keywordMatch(journal, journal[sample(nrow(journal)),2], one.to.one = TRUE)
}
\keyword{by}
\keyword{cosine}
\keyword{keyword,}
\keyword{match}
\keyword{similarity}

